03-data-warehouse/requirements.txt
google-cloud-storage==2.14.0
pandas==2.2.0
pyarrow==15.0.0